Backend/requirements.txt
alembic==1.5.8
bcrypt==3.2.0
boto3==1.17.64
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
marshmallow==3.11.1
psycopg2==2.8.6
PyJWT==2.1.0
python-dotenv==0.17.1
SQLAlchemy==1.4.12
Werkzeug==1.0.1
pytest==6.2.4
pytest-cov==2.11.1